requirements.txt
python-telegram-bot
ultralytics
google-generativeai
python-dotenv
Pillow